Implement consolidateunspent wizard
This implements a three step wizard that leads the user through
the process of selecting inputs, selecting a destination, and
then reviewing the overall transaction.

Select Inputs:

The select inputs screen uses similar code to that in coincontroldialog
to support the consolidate button there. Pointers to the coincontrol
data structures constructed in sendcoinsdialog are passed into
both coincontrol and the consolidateunspentwizard to faciliate using
the underlying machinery in a unified manner. This is possible because
both coincontrol and consolidateunspendwizard are called with the
sendcoinsdialog and are modal.

Note that there is a stop sign and the next button is disabled if more than
600 outputs are selected. The next button is also disabled if less than 2
outputs are selected (as it makes no sense to consolidate when there is no
consolidation achievable based on the selection). If a filter operation
is applied based on the filter criteria, and that criteria would result in
more than 600 inputs being selected, the selection is reduced to 600 inputs
and a warning triangle is presented.

Select Destination Address:

If all of the inputs selected in the prior page are from one address, then
that address will already be preselected (but allow the opportunity for it
to be changed by the user prior to pressing next). If the inputs selected
correspond to more than one address, an address will NOT be pre-selected,
and the user will be required to pick an address to proceed. The next button
will be disabled until a valid address is selected for the destination.

Confirmation:

The final screen presents the details of the intended transaction for review
by the user. When the "Finish" button is pressed, the transaction to send
is filled in in the send dialog sreen, and the user can review the details
again if desired and press the send button to send.